[refec] Map관련 코드 기능별로 분리
