+++ front/src/App.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import "./App.css";
 import Layout from "./components/Layout/Layout";
-import { ServerMap } from "./components/Phaser/ServerMap";
+import { ServerMap } from "./components/Map/ServerMap";
 import Home from "./pages/Home";
 import Join from "./pages/Join";
 import Login from "./pages/Login";
