+++ front/src/components/Map/ServerMap.tsx
