+++ front/src/components/Map/ServerMap/MainMap.ts
@@ -1,0 +1,91 @@
+import { Scene } from "phaser";
+import { Layers } from "../../../types/server";
+import { charanterAnimation, createCharacter } from "./Character";
+import { createLayers } from "./Layers";
+import { addCollider, debugCollision } from "./Debug";
+
+export class MainMap extends Scene {
+  private mapInstance!: Phaser.Game;
+  private character!: Phaser.Physics.Arcade.Sprite;
+  private keyboards!: Phaser.Types.Input.Keyboard.CursorKeys;
+  private isDragging: boolean = false;
+  private dragStartX: number = 0;
+  private dragStartY: number = 0;
+
+  constructor() {
+    super("MainMap");
+  }
+
+  init(data: { mapInstance: Phaser.Game }) {
+    this.mapInstance = data.mapInstance;
+  }
+
+  create() {
+    // preload에서 tilemapTiledJSON에서 지정한 string key와 매치시켜 map으로 지정
+    const map = this.make.tilemap({ key: "firstMap" });
+
+    // 타일 레이어 생성
+    const layers = createLayers(map, this);
+
+    // 타일 레이어가 생성되었을 경우에 debugCollision 함수 호출
+    if (layers) {
+      debugCollision(this, layers);
+    } else {
+      console.log("레이어가 생성되지 않았습니다.");
+    }
+
+    // 캐릭터 택스처 파일 로드 완료 여부에 따른 함수 호출
+    this.load.on("complete", () => {
+      try {
+        this.character = createCharacter(this, 300, 300);
+        charanterAnimation(this);
+
+        this.keyboards = this.input.keyboard?.createCursorKeys()!; // 방향키로 캐리터 조정
+        this.cameras.main.startFollow(this.character); // 캐릭터의 움직을 따라 카메라 움직임
+
+        // 캐릭터가 접근할 수 없는 곳 블로킹
+        if (layers) {
+          addCollider(this, this.character, layers);
+        } else {
+          console.log("레이어가 생성되지 않았습니다.");
+        }
+      } catch (error) {
+        console.error(error);
+      }
+    });
+
+    this.load.start();
+
+    const mapWidth = map.widthInPixels;
+    const mapHeight = map.heightInPixels;
+    this.cameras.main.setBounds(0, 0, mapWidth, mapHeight);
+    this.physics.world.setBounds(0, 0, mapWidth, mapHeight);
+
+    return layers;
+  }
+
+  update() {
+    if (!this.character) {
+      // 만약 캐리터가 로드되지 않았다면 아무것도 반환하지 않는다.
+      return;
+    }
+
+    this.character.setVelocity(0);
+
+    if (this.keyboards.down.isDown) {
+      this.character.setVelocityY(100);
+      this.character.anims.play("basic_character_move_down", true);
+    } else if (this.keyboards.up.isDown) {
+      this.character.setVelocityY(-100);
+      this.character.anims.play("basic_character_move_up", true);
+    } else if (this.keyboards.left.isDown) {
+      this.character.setVelocityX(-100);
+      this.character.anims.play("basic_character_move_left", true);
+    } else if (this.keyboards.right.isDown) {
+      this.character.setVelocityX(100);
+      this.character.anims.play("basic_character_move_right", true);
+    } else {
+      this.character.anims.stop();
+    }
+  }
+}
