+++ front/src/components/Map/ServerMap/Preloader.ts
